Fix i18n load path for use CARGO_MANIFEST_DIR.
And complete intergration test.
