+++ README.md
@@ -8,13 +8,22 @@ The API of this crate is inspired by [ruby-i18n](https://github.com/ruby-i18n/i1
 
 ## Usage
 
-### Preparing the Localized Mappings
+Add crate dependencies in your Cargo.toml:
+
+```toml
+[dependencies]
+lazy_static = "1.4.0"
+rust-i18n = "0"
+```
+
+Load macro in your `lib.rs`
 
 ```rs
 // Load I18n macro, for allow you use `t!` macro in anywhere.
 #[macro_use]
 extern crate rust_i18n;
 
+// Init translations for current crate.
 i18n!("./locales");
 ```
