--- examples/bar/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-edition = "2021"
-name = "bar"
-version = "0.1.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-foo = {path = "../foo"}
-lazy_static = "1.4.0"
-rust-i18n = {path = "../../../rust-i18n"}
