--- examples/bar/locales/en.yml
@@ -1,2 +1,0 @@
-en:
-  hello: bar - Hello
