+++ examples/foo/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 edition = "2021"
 name = "foo"
+private = true
 version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
