+++ tests/locales/de.yml
@@ -1,4 +1,4 @@
 de:
-  hello: Foo - Hallo Welt!
+  hello: Bar - Hallo Welt!
   messages:
     hello: Hallo, %{name}!
