+++ tests/locales/en.yml
@@ -1,0 +1,15 @@
+en:
+  hello: Bar - Hello, World!
+  custom:
+    greeting: Hello, %{name}!!!
+  a:
+    very:
+      nested:
+        message: "Hello, %{name}. Your message is: %{msg}"
+  messages:
+    zero: You have no messages.
+    one: You have one message.
+    other: You have %{count} messages.
+    hello: Hello, %{name}!
+  missing:
+    default: Sorry, that translation doesn't exist.
